+++ http-service/src/main/java/net/runelite/http/service/item/RSPrices.java
@@ -37,5 +37,5 @@ class RSPrices
 	/**
 	 * item id to price
 	 */
-	Map<Integer, Integer> prices;
+	Map<Integer, Long> prices;
 }
